requirements.txt
streamlit
google-generativeai
folium
requests
streamlit-folium